Remove add class, add `last_version_number` method in `inject.rb`
