+++ lib/bundler/cli/inject.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
+require "bundler/cli/common"
 module Bundler
   class CLI::Inject
     attr_reader :options, :name, :version, :gems
     def initialize(options, name, version, gems)
       @options = options
       @name = name
-      @version = version
+      @version = version || last_version_number
       @gems = gems
     end
 
@@ -28,6 +29,17 @@ module Bundler
       else
         Bundler.ui.confirm "All injected gems were already present in the Gemfile"
       end
+    end
+
+  private
+
+    def last_version_number
+      definition = Bundler.definition(true)
+      definition.resolve_remotely!
+      specs = definition.index[name].sort_by(&:version)
+      spec = specs.delete_if {|b| b.respond_to?(:version) && b.version.prerelease? }
+      spec = specs.last
+      spec.version.to_s
     end
   end
 end
